AIDraw/requirements.txt
numpy~=1.23.4
scipy~=1.9.3
pytorch-lightning==1.7.6
opencv-python~=4.6.0.66
pyyaml~=6.0
omegaconf~=2.2.3
tqdm~=4.64.1
einops==0.5.0
diffusers~=0.18.0
clip==0.2.0
kornia==0.6.8
transformers>=4.26
controlnet_aux~=0.0.6
matplotlib~=3.7.2
insightface~=0.7.3
onnxruntime~=1.12.0
safetensors~=0.3.1
timm~=0.9.2
compel